+++ examples/next-openai/app/function-calling/page.tsx
@@ -14,8 +14,10 @@ export default function Chat() {
         const parsedFunctionCallArguments: { code: string } = JSON.parse(
           functionCall.arguments,
         );
+
         // WARNING: Do NOT do this in real-world applications!
         eval(parsedFunctionCallArguments.code);
+
         const functionResponse = {
           messages: [
             ...chatMessages,
@@ -27,6 +29,7 @@ export default function Chat() {
             },
           ],
         };
+
         return functionResponse;
       }
     }
@@ -42,7 +45,9 @@ export default function Chat() {
     system: 'red',
     user: 'black',
     function: 'blue',
+    tool: 'purple',
     assistant: 'green',
+    data: 'orange',
   };
 
   return (