+++ packages/core/streams/anthropic-stream.test.ts
@@ -1,28 +1,48 @@
+import Anthropic from '@anthropic-ai/sdk';
 import {
   AnthropicStream,
   StreamingTextResponse,
   experimental_StreamData,
 } from '.';
+import {
+  anthropicMessageChunks,
+  anthropicCompletionChunks,
+} from '../tests/snapshots/anthropic';
 import { readAllChunks } from '../tests/utils/mock-client';
-import { setup } from '../tests/utils/mock-service';
-import Anthropic from '@anthropic-ai/sdk';
+import { DEFAULT_TEST_URL, createMockServer } from '../tests/utils/mock-server';
 
-describe('AnthropicStream', () => {
-  let server: ReturnType<typeof setup>;
-  beforeAll(() => {
-    server = setup(3035);
-  });
-  afterAll(async () => server.teardown());
+export const MESSAGE_URL = 'http://localhost:3030/messages';
 
+const server = createMockServer([
+  {
+    url: DEFAULT_TEST_URL,
+    chunks: anthropicCompletionChunks,
+    formatChunk: chunk =>
+      `event: completion\ndata: ${JSON.stringify(chunk)}\n\n`,
+  },
+  {
+    url: MESSAGE_URL,
+    chunks: anthropicMessageChunks,
+    formatChunk: chunk => chunk,
+  },
+]);
+
+beforeAll(() => {
+  server.listen();
+});
+
+afterEach(() => {
+  server.resetHandlers();
+});
+
+afterAll(() => {
+  server.close();
+});
+
+describe('Anthropic completion', () => {
   it('should be able to parse SSE and receive the streamed response', async () => {
     const anthropic = new Anthropic({
-      fetch: () =>
-        fetch(server.api, {
-          headers: {
-            'x-mock-service': 'anthropic',
-            'x-mock-type': 'chat',
-          },
-        }),
+      fetch: () => fetch(DEFAULT_TEST_URL),
       apiKey: 'sk-doesnt-matter',
     });
 
@@ -48,13 +68,7 @@ describe('AnthropicStream', () => {
   describe('StreamData protocol', () => {
     it('should send text', async () => {
       const anthropic = new Anthropic({
-        fetch: () =>
-          fetch(server.api, {
-            headers: {
-              'x-mock-service': 'anthropic',
-              'x-mock-type': 'chat',
-            },
-          }),
+        fetch: () => fetch(DEFAULT_TEST_URL),
         apiKey: 'sk-doesnt-matter',
       });
 
@@ -86,13 +100,7 @@ describe('AnthropicStream', () => {
 
     it('should send text and data', async () => {
       const anthropic = new Anthropic({
-        fetch: () =>
-          fetch(server.api, {
-            headers: {
-              'x-mock-service': 'anthropic',
-              'x-mock-type': 'chat',
-            },
-          }),
+        fetch: () => fetch(DEFAULT_TEST_URL),
         apiKey: 'sk-doesnt-matter',
       });
 
@@ -119,6 +127,102 @@ describe('AnthropicStream', () => {
       expect(await readAllChunks(response)).toEqual([
         '2:[{"t1":"v1"}]\n',
         '0:" Hello"\n',
+        '0:","\n',
+        '0:" world"\n',
+        '0:"."\n',
+      ]);
+    });
+  });
+});
+
+describe('Anthropic message', () => {
+  it('should be able to parse SSE and receive the streamed response', async () => {
+    const anthropic = new Anthropic({
+      fetch: () => fetch(MESSAGE_URL),
+      apiKey: 'sk-doesnt-matter',
+    });
+
+    const anthropicResponse = await anthropic.beta.messages.create({
+      messages: [{ role: 'user', content: 'Hello' }],
+      model: 'claude-2.1',
+      stream: true,
+      max_tokens: 300,
+    });
+
+    const stream = AnthropicStream(anthropicResponse);
+
+    const response = new StreamingTextResponse(stream);
+
+    expect(await readAllChunks(response)).toEqual([
+      'Hello',
+      ',',
+      ' world',
+      '.',
+    ]);
+  });
+
+  describe('StreamData protocol', () => {
+    it('should send text', async () => {
+      const anthropic = new Anthropic({
+        fetch: () => fetch(MESSAGE_URL),
+        apiKey: 'sk-doesnt-matter',
+      });
+
+      const data = new experimental_StreamData();
+
+      const anthropicResponse = await anthropic.beta.messages.create({
+        messages: [{ role: 'user', content: 'Hello' }],
+        model: 'claude-2.1',
+        stream: true,
+        max_tokens: 300,
+      });
+
+      const stream = AnthropicStream(anthropicResponse, {
+        onFinal() {
+          data.close();
+        },
+        experimental_streamData: true,
+      });
+
+      const response = new StreamingTextResponse(stream, {}, data);
+
+      expect(await readAllChunks(response)).toEqual([
+        '0:"Hello"\n',
+        '0:","\n',
+        '0:" world"\n',
+        '0:"."\n',
+      ]);
+    });
+
+    it('should send text and data', async () => {
+      const anthropic = new Anthropic({
+        fetch: () => fetch(MESSAGE_URL),
+        apiKey: 'sk-doesnt-matter',
+      });
+
+      const data = new experimental_StreamData();
+
+      data.append({ t1: 'v1' });
+
+      const anthropicResponse = await anthropic.beta.messages.create({
+        messages: [{ role: 'user', content: 'Hello' }],
+        model: 'claude-2.1',
+        stream: true,
+        max_tokens: 300,
+      });
+
+      const stream = AnthropicStream(anthropicResponse, {
+        onFinal() {
+          data.close();
+        },
+        experimental_streamData: true,
+      });
+
+      const response = new StreamingTextResponse(stream, {}, data);
+
+      expect(await readAllChunks(response)).toEqual([
+        '2:[{"t1":"v1"}]\n',
+        '0:"Hello"\n',
         '0:","\n',
         '0:" world"\n',
         '0:"."\n',