+++ packages/core/streams/huggingface-stream.test.ts
@@ -4,17 +4,32 @@ import {
   StreamingTextResponse,
   experimental_StreamData,
 } from '.';
+import { huggingfaceChunks } from '../tests/snapshots/huggingface';
 import { createClient } from '../tests/utils/mock-client';
-import { setup } from '../tests/utils/mock-service';
+import { DEFAULT_TEST_URL, createMockServer } from '../tests/utils/mock-server';
+
+const server = createMockServer([
+  {
+    url: DEFAULT_TEST_URL,
+    chunks: huggingfaceChunks,
+    formatChunk: chunk => `data: ${JSON.stringify(chunk)}\n\n`,
+  },
+]);
 
 describe('HuggingFace stream', () => {
   const Hf = new HfInference();
 
-  let server: ReturnType<typeof setup>;
   beforeAll(() => {
-    server = setup(3033);
+    server.listen();
   });
-  afterAll(async () => server.teardown());
+
+  afterEach(() => {
+    server.resetHandlers();
+  });
+
+  afterAll(() => {
+    server.close();
+  });
 
   function readAllChunks(response: Response) {
     return createClient(response).readAll();
@@ -24,16 +39,7 @@ describe('HuggingFace stream', () => {
     const stream = HuggingFaceStream(
       Hf.textGenerationStream(
         { model: 'model', inputs: '' },
-        {
-          fetch() {
-            return fetch(server.api, {
-              headers: {
-                'x-mock-service': 'huggingface',
-                'x-mock-type': 'chat',
-              },
-            });
-          },
-        },
+        { fetch: () => fetch(DEFAULT_TEST_URL) },
       ),
     );
 
@@ -54,16 +60,7 @@ describe('HuggingFace stream', () => {
       const stream = HuggingFaceStream(
         Hf.textGenerationStream(
           { model: 'model', inputs: '' },
-          {
-            fetch() {
-              return fetch(server.api, {
-                headers: {
-                  'x-mock-service': 'huggingface',
-                  'x-mock-type': 'chat',
-                },
-              });
-            },
-          },
+          { fetch: () => fetch(DEFAULT_TEST_URL) },
         ),
         {
           onFinal() {
@@ -91,16 +88,7 @@ describe('HuggingFace stream', () => {
       const stream = HuggingFaceStream(
         Hf.textGenerationStream(
           { model: 'model', inputs: '' },
-          {
-            fetch() {
-              return fetch(server.api, {
-                headers: {
-                  'x-mock-service': 'huggingface',
-                  'x-mock-type': 'chat',
-                },
-              });
-            },
-          },
+          { fetch: () => fetch(DEFAULT_TEST_URL) },
         ),
         {
           onFinal() {